+++ src/bittytax/conv/parsers/__init__.py
@@ -16,6 +16,7 @@ from . import (
     blockfi,
     blockpit,
     blockscout,
+    blofin,
     bnktothefuture,
     bybit,
     cashapp,